Romy - added commands and shi
