+++ src/main/java/edu/greenblitz/robotName/commands/shootingPosition/FixShooterAngle.java
@@ -1,0 +1,18 @@
+package edu.greenblitz.robotName.commands.shootingPosition;
+
+import edu.greenblitz.robotName.utils.GBCommand;
+
+public class FixShooterAngle extends GBCommand {
+//    private ShootingState state;
+//    private Pivot pivot;
+//
+//    public FixShooterAngle() {
+//        state = ShootingState.getInstance();
+//        pivot = Pivot.getInstance();
+//        require(pivot);
+//    }
+//    @Override
+//    public void execute() {
+//        new Pivot.getGoToAngle()
+//    }
+}
