+++ src/main/java/edu/greenblitz/robotName/commands/shootingPosition/GoToShootingPosition.java
@@ -1,0 +1,7 @@
+package edu.greenblitz.robotName.commands.shootingPosition;
+
+import edu.greenblitz.robotName.utils.GBCommand;
+
+public class GoToShootingPosition extends GBCommand {
+
+}
