+++ src/main/java/edu/greenblitz/robotName/commands/shootingPosition/ShootingPositionConstants.java
@@ -1,0 +1,11 @@
+package edu.greenblitz.robotName.commands.shootingPosition;
+
+import java.awt.geom.Ellipse2D;
+
+public class ShootingPositionConstants {
+    public static final int X = 69;
+    public static final int Y = 69;
+    public static final int WIDTH = 69;
+    public static final int HIGHT = 69;
+    public static final Ellipse2D CURVE = new Ellipse2D.Double(X, Y, WIDTH, HIGHT);
+}
