+++ src/main/java/edu/greenblitz/robotName/utils/GBMath.java
@@ -1,7 +1,5 @@
 package edu.greenblitz.robotName.utils;
 
-import edu.greenblitz.robotName.FieldConstants;
-import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.geometry.Translation3d;
